refactor: rename and merge type modules per maintainer review
Rename the modules introduced in PR #7 so each ends with `_types`, and
merge two pairs:

  org_and_tool_extras           -> extra_types
  protocol_message_types        -> message_types
  supabase_database             -> database_types
  webhook_types + webhook_error -> whatsapp_webhook_payload_types
  whatsapp_incoming_message_types -> whatsapp_webhook_message_types
  whatsapp_cloud_outgoing_types merged into whatsapp_endpoint_types

Imports inside `_shared/types/` updated accordingly; the `supabase.ts`
barrel re-exports the new paths so callsites are unchanged.
