--- supabase/functions/_shared/types/whatsapp_cloud_outgoing_types.ts
@@ -1,56 +1,0 @@
-//===================================
-// Outgoing message, as sent to the WhatsApp Cloud API
-// (partial types; combined into EndpointMessage in whatsapp_endpoint_types.ts)
-//===================================
-
-export type OutgoingContextInfo = {
-  context?: { message_id: string };
-};
-
-// Text based
-
-export type OutgoingText = {
-  type: "text";
-  text: {
-    body: string;
-    preview_url?: boolean;
-  };
-};
-
-export type OutgoingReaction = {
-  type: "reaction";
-  reaction: {
-    emoji: string;
-    message_id: string;
-  };
-};
-
-// File based
-
-export type OutgoingAudio = {
-  type: "audio";
-  audio: ({ id: string } | { link: string }) & { voice?: boolean };
-};
-
-export type OutgoingImage = {
-  type: "image";
-  image: ({ id: string } | { link: string }) & { caption?: string };
-};
-
-export type OutgoingVideo = {
-  type: "video";
-  video: ({ id: string } | { link: string }) & { caption?: string };
-};
-
-export type OutgoingDocument = {
-  type: "document";
-  document: ({ id: string } | { link: string }) & {
-    caption?: string;
-    filename?: string;
-  };
-};
-
-export type OutgoingSticker = {
-  type: "sticker";
-  sticker: { id: string } | { link: string };
-};
