+++ supabase/functions/_shared/types/whatsapp_endpoint_types.ts
@@ -1,18 +1,64 @@
 import type {
   ContactsMessage,
   LocationMessage,
-} from "./whatsapp_incoming_message_types.ts";
+} from "./whatsapp_webhook_message_types.ts";
 import type { TemplateMessage } from "./whatsapp_template_types.ts";
-import type {
-  OutgoingAudio,
-  OutgoingContextInfo,
-  OutgoingDocument,
-  OutgoingImage,
-  OutgoingReaction,
-  OutgoingSticker,
-  OutgoingText,
-  OutgoingVideo,
-} from "./whatsapp_cloud_outgoing_types.ts";
+
+//===================================
+// Outgoing message components, as sent to the WhatsApp Cloud API
+//===================================
+
+export type OutgoingContextInfo = {
+  context?: { message_id: string };
+};
+
+// Text based
+
+export type OutgoingText = {
+  type: "text";
+  text: {
+    body: string;
+    preview_url?: boolean;
+  };
+};
+
+export type OutgoingReaction = {
+  type: "reaction";
+  reaction: {
+    emoji: string;
+    message_id: string;
+  };
+};
+
+// File based
+
+export type OutgoingAudio = {
+  type: "audio";
+  audio: ({ id: string } | { link: string }) & { voice?: boolean };
+};
+
+export type OutgoingImage = {
+  type: "image";
+  image: ({ id: string } | { link: string }) & { caption?: string };
+};
+
+export type OutgoingVideo = {
+  type: "video";
+  video: ({ id: string } | { link: string }) & { caption?: string };
+};
+
+export type OutgoingDocument = {
+  type: "document";
+  document: ({ id: string } | { link: string }) & {
+    caption?: string;
+    filename?: string;
+  };
+};
+
+export type OutgoingSticker = {
+  type: "sticker";
+  sticker: { id: string } | { link: string };
+};
 
 //===================================
 // Endpoint message, as sent to the WhatsApp endpoint
